Added state and maintenance checks
