+++ frontend/src/pages/Risk/Risk.tsx
@@ -1,15 +1,6 @@
 import React, { useCallback, useState, useEffect, useMemo } from 'react';
 import classes from './Risk.module.scss';
-import {
-  Box,
-  Card,
-  CardContent,
-  Grid,
-  Paper,
-  Typography,
-  Button
-} from '@mui/material';
-import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
+import { Box, Grid, Paper } from '@mui/material';
 import VulnerabilityCard from './VulnerabilityCard';
 import TopVulnerablePorts from './TopVulnerablePorts';
 import TopVulnerableDomains from './TopVulnerableDomains';
@@ -28,7 +19,6 @@ import {
 } from 'react-simple-maps';
 import { scaleLinear } from 'd3-scale';
 import { Stats, Vulnerability } from 'types';
-import { UpdateStateForm } from 'components/Register';
 import {
   ORGANIZATION_FILTER_KEY,
   OrganizationShallow,
@@ -40,7 +30,6 @@ import { useLocation } from 'react-router-dom';
 import { useUserTypeFilters } from 'hooks/useUserTypeFilters';
 import { useStaticsContext } from 'context/StaticsContext';
 import { useUserLevel } from 'hooks/useUserLevel';
-import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
 
 export interface Point {
   id: string;
@@ -75,18 +64,9 @@ const Risk: React.FC<ContextType> = ({
   search_term,
   setSearchTerm
 }) => {
-  const {
-    showMaps,
-    user,
-    apiPost,
-    apiGet,
-    logout,
-    userMustSign,
-    isLoggingOut
-  } = useAuthContext();
+  const { showMaps, user, apiPost } = useAuthContext();
 
   const [stats, setStats] = useState<Stats | undefined>(undefined);
-  const [isUpdateStateFormOpen, setIsUpdateStateFormOpen] = useState(false);
 
   const RiskRoot = RiskStyles.RiskRoot;
   const { cardRoot, content, contentWrapper, header, panel } =
@@ -165,77 +145,9 @@ const Risk: React.FC<ContextType> = ({
     [riskFilters]
   );
 
-  const [isLoginBlockedDialogOpen, setIsLoginBlockedDialogOpen] =
-    useState(false);
-  const [maintenanceNotification, setMaintenanceNotification] =
-    useState<any>(null);
-
   useEffect(() => {
     fetchStats();
   }, [fetchStats, riskFilters]);
-
-  useEffect(() => {
-    if (!isLoggingOut && user) {
-      if (
-        (!user.state || user.state === '') &&
-        !localStorage.getItem('user_state')
-      ) {
-        setIsUpdateStateFormOpen(true);
-      }
-    }
-  }, [user, isLoggingOut]);
-
-  useEffect(() => {
-    const fetchAndCheckMaintenance = async () => {
-      // TODO: Some login blocking logic is a duplicate of backend
-      // checks to meet "waiting room" needs to allow controlled logins
-      // for populating new user state and terms agreement before blocking
-      // for pre-release. Standardize this once this is no longer needed.
-      if (
-        user &&
-        !user.invite_pending &&
-        user.state &&
-        user.date_accepted_terms &&
-        !isLoginBlockedDialogOpen
-      ) {
-        // Active Notifications
-        const notifications = await apiGet('/notifications');
-        const active = notifications.find(
-          (n: any) =>
-            n.status === 'active' &&
-            n.maintenance_type === 'major' &&
-            new Date(n.start_datetime) <= new Date() &&
-            new Date(n.end_datetime) >= new Date()
-        );
-        // Set non-blocking userTypes (additional check)
-        const nonBlockingUserTypes = ['globalAdmin', 'regionalAdmin'];
-        if (active && !nonBlockingUserTypes.includes(user.user_type)) {
-          setMaintenanceNotification(active);
-          setIsLoginBlockedDialogOpen(true);
-        }
-      }
-    };
-
-    fetchAndCheckMaintenance();
-  }, [apiGet, isLoginBlockedDialogOpen, user, userMustSign]);
-
-  useEffect(() => {
-    const handleMaintenanceBlocked = (e: any) => {
-      if (e.detail?.message) {
-        setMaintenanceNotification({ message: e.detail.message });
-        setIsLoginBlockedDialogOpen(true);
-      }
-    };
-
-    window.addEventListener('maintenance-blocked', handleMaintenanceBlocked);
-
-    return () => {
-      window.removeEventListener(
-        'maintenance-blocked',
-        handleMaintenanceBlocked
-      );
-    };
-  }, []);
 
   useEffect(() => {
     filters.forEach((filter) => {
@@ -367,96 +279,6 @@ const Risk: React.FC<ContextType> = ({
         sev.sevList.includes(i.id.split('|')[1])
       );
     }
-  }
-
-  if (isUpdateStateFormOpen) {
-    return (
-      <UpdateStateForm
-        open={isUpdateStateFormOpen}
-        user_id={user?.id ?? ''}
-        onClose={async () => {
-          setIsUpdateStateFormOpen(false);
-
-          // Re-fetch updated user to prevent false popup
-          const updatedUser = await apiGet('/users/me');
-          if (updatedUser?.state && user?.user_type !== 'globalAdmin') {
-            const notifications = await apiGet('/notifications');
-            const active = notifications.find(
-              (n: any) =>
-                n.status === 'active' &&
-                n.maintenance_type === 'major' &&
-                new Date(n.start_datetime) <= new Date() &&
-                new Date(n.end_datetime) >= new Date()
-            );
-            if (active && updatedUser.user_type !== 'globalAdmin') {
-              setMaintenanceNotification(active);
-              setIsLoginBlockedDialogOpen(true);
-            }
-          }
-        }}
-      />
-    );
-  }
-
-  if (isLoginBlockedDialogOpen && maintenanceNotification) {
-    return (
-      <LoginBlockedDialog
-        open={isLoginBlockedDialogOpen}
-        message={maintenanceNotification.message}
-        onClose={() => {
-          setIsLoginBlockedDialogOpen(false);
-          logout();
-        }}
-      />
-    );
-  }
-
-  if (user?.invite_pending) {
-    return (
-      <div
-        style={{
-          position: 'fixed',
-          top: 0,
-          left: 0,
-          width: '100vw',
-          height: '100vh',
-          background: 'rgba(0,0,0,0.4)',
-          zIndex: 1300,
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center'
-        }}
-      >
-        <Card style={{ maxWidth: 400, textAlign: 'center' }}>
-          <CardContent>
-            <CheckCircleOutlineIcon
-              sx={{
-                fontSize: 48,
-                color: (theme) => theme.palette.primary.dark,
-                mb: 1
-              }}
-            />
-            <Typography variant="body1" sx={{ mb: 2 }}>
-              Request sent! We will notify you by email once your account is
-              approved
-            </Typography>
-            <Button
-              variant="contained"
-              sx={{
-                backgroundColor: (theme) => theme.palette.primary.dark,
-                color: '#fff',
-                '&:hover': {
-                  backgroundColor: (theme) => theme.palette.primary.main
-                }
-              }}
-              onClick={logout}
-            >
-              Logout
-            </Button>
-          </CardContent>
-        </Card>
-      </div>
-    );
   }
 
   return (
