+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -5,11 +5,18 @@ import {
   BoxProps,
   Grid,
   CircularProgress,
-  Stack
+  Stack,
+  Card,
+  CardContent,
+  Typography,
+  Button
 } from '@mui/material';
+import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline';
 import KeyMetrics from './KeyMetrics';
 import ScanningSummary from './ScanningSummary';
 import InfoLabel from 'components/Dashboard/InfoLabel';
+import { LoginBlockedDialog } from 'components/LoginBlockedDialog';
+import { UpdateStateForm } from 'components/Register';
 import GraphChip from 'components/Dashboard/GraphChip';
 import {
   // StatsTrendsRawData,
@@ -63,7 +70,16 @@ export const transformScanData = (rawData: any) => {
 };
 
 const VulnerabilityScan: React.FC = () => {
-  const { apiPost, currentOrganization } = useAuthContext();
+  const {
+    apiPost,
+    currentOrganization,
+    user,
+    apiGet,
+    userMustSign,
+    logout,
+    isLoggingOut
+  } = useAuthContext();
+  const [isUpdateStateFormOpen, setIsUpdateStateFormOpen] = useState(false);
   // const [statsTrendsRawData, setStatsTrendsRawData] = useState<
   //   StatsTrendsRawData | undefined
   // >(undefined);
@@ -93,6 +109,74 @@ const VulnerabilityScan: React.FC = () => {
     setSeverityByProminenceVulnType(label);
     setSeverityByProminenceActiveLabel(label);
   };
+
+  const [isLoginBlockedDialogOpen, setIsLoginBlockedDialogOpen] =
+    useState(false);
+  const [maintenanceNotification, setMaintenanceNotification] =
+    useState<any>(null);
+
+  useEffect(() => {
+    if (!isLoggingOut && user) {
+      if (
+        (!user.state || user.state === '') &&
+        !localStorage.getItem('user_state')
+      ) {
+        setIsUpdateStateFormOpen(true);
+      }
+    }
+  }, [user, isLoggingOut]);
+
+  useEffect(() => {
+    const fetchAndCheckMaintenance = async () => {
+      // TODO: Some login blocking logic is a duplicate of backend
+      // checks to meet "waiting room" needs to allow controlled logins
+      // for populating new user state and terms agreement before blocking
+      // for pre-release. Standardize this once this is no longer needed.
+      if (
+        user &&
+        !user.invite_pending &&
+        user.state &&
+        user.date_accepted_terms &&
+        !isLoginBlockedDialogOpen
+      ) {
+        // Active Notifications
+        const notifications = await apiGet('/notifications');
+        const active = notifications.find(
+          (n: any) =>
+            n.status === 'active' &&
+            n.maintenance_type === 'major' &&
+            new Date(n.start_datetime) <= new Date() &&
+            new Date(n.end_datetime) >= new Date()
+        );
+        // Set non-blocking userTypes (additional check)
+        const nonBlockingUserTypes = ['globalAdmin', 'regionalAdmin'];
+        if (active && !nonBlockingUserTypes.includes(user.user_type)) {
+          setMaintenanceNotification(active);
+          setIsLoginBlockedDialogOpen(true);
+        }
+      }
+    };
+
+    fetchAndCheckMaintenance();
+  }, [apiGet, isLoginBlockedDialogOpen, user, userMustSign]);
+
+  useEffect(() => {
+    const handleMaintenanceBlocked = (e: any) => {
+      if (e.detail?.message) {
+        setMaintenanceNotification({ message: e.detail.message });
+        setIsLoginBlockedDialogOpen(true);
+      }
+    };
+
+    window.addEventListener('maintenance-blocked', handleMaintenanceBlocked);
+
+    return () => {
+      window.removeEventListener(
+        'maintenance-blocked',
+        handleMaintenanceBlocked
+      );
+    };
+  }, []);
 
   //Dynamically generate, format, and sort severity levels for Top Vuln Hosts' GraphChips from the data
   const presentSeverityLevels = Array.from(
@@ -213,6 +297,96 @@ const VulnerabilityScan: React.FC = () => {
       </Grid>
     </Box>
   );
+
+  if (isUpdateStateFormOpen) {
+    return (
+      <UpdateStateForm
+        open={isUpdateStateFormOpen}
+        user_id={user?.id ?? ''}
+        onClose={async () => {
+          setIsUpdateStateFormOpen(false);
+
+          // Re-fetch updated user to prevent false popup
+          const updatedUser = await apiGet('/users/me');
+          if (updatedUser?.state && user?.user_type !== 'globalAdmin') {
+            const notifications = await apiGet('/notifications');
+            const active = notifications.find(
+              (n: any) =>
+                n.status === 'active' &&
+                n.maintenance_type === 'major' &&
+                new Date(n.start_datetime) <= new Date() &&
+                new Date(n.end_datetime) >= new Date()
+            );
+            if (active && updatedUser.user_type !== 'globalAdmin') {
+              setMaintenanceNotification(active);
+              setIsLoginBlockedDialogOpen(true);
+            }
+          }
+        }}
+      />
+    );
+  }
+
+  if (isLoginBlockedDialogOpen && maintenanceNotification) {
+    return (
+      <LoginBlockedDialog
+        open={isLoginBlockedDialogOpen}
+        message={maintenanceNotification.message}
+        onClose={() => {
+          setIsLoginBlockedDialogOpen(false);
+          logout();
+        }}
+      />
+    );
+  }
+
+  if (user?.invite_pending) {
+    return (
+      <div
+        style={{
+          position: 'fixed',
+          top: 0,
+          left: 0,
+          width: '100vw',
+          height: '100vh',
+          background: 'rgba(0,0,0,0.4)',
+          zIndex: 1300,
+          display: 'flex',
+          justifyContent: 'center',
+          alignItems: 'center'
+        }}
+      >
+        <Card style={{ maxWidth: 400, textAlign: 'center' }}>
+          <CardContent>
+            <CheckCircleOutlineIcon
+              sx={{
+                fontSize: 48,
+                color: (theme) => theme.palette.primary.dark,
+                mb: 1
+              }}
+            />
+            <Typography variant="body1" sx={{ mb: 2 }}>
+              Request sent! We will notify you by email once your account is
+              approved
+            </Typography>
+            <Button
+              variant="contained"
+              sx={{
+                backgroundColor: (theme) => theme.palette.primary.dark,
+                color: '#fff',
+                '&:hover': {
+                  backgroundColor: (theme) => theme.palette.primary.main
+                }
+              }}
+              onClick={logout}
+            >
+              Logout
+            </Button>
+          </CardContent>
+        </Card>
+      </div>
+    );
+  }
 
   if (loading) {
     return (
